warehouse/requirements.txt
asgiref==3.9.1
colorama==0.4.6
Django==5.2.5
pillow==11.3.0
python-barcode==0.15.1
qrcode==8.2
sqlparse==0.5.3
tzdata==2025.2
gunicorn>=21.2
djangorestframework>=3.15.0
django-extensions
drf-spectacular